feat(testing): add Vitest browser mode with Playwright Firefox
Split AuthenticationForm into LoginForm, RegisterForm, and AuthMessage
components. Set up Vitest browser mode using Playwright/Firefox with
vitest-browser-react for component testing. Add Login.test.tsx and
update Header.test.tsx to use browser APIs (page, expect.element).
Configure CI to install Playwright Firefox before running tests.

- Split AuthenticationForm.tsx into auth/LoginForm.tsx,
  auth/RegisterForm.tsx, and auth/AuthMessage.tsx
- Configure vite.config.ts with playwright() provider and Firefox instance
- Install @vitest/browser, @vitest/browser-playwright, vitest-browser-react
- Update Login.test.tsx: validation tests + fetch spy for API call
- Update Header.test.tsx for Vitest browser mode APIs
- Add Playwright Firefox install step in CI workflow
- Remove obsolete test.tsx
