+++ .github/workflows/deployTSH.yml
@@ -34,8 +34,11 @@ jobs:
         env:
           VITE_API_BASE_URL: /api
 
+      - name: Install Playwright Firefox
+        run: npx playwright install firefox --with-deps
+
       - name: Run tests
-        run: npm test
+        run: npm run test -- --run
 
       - name: Upload build artifacts
         uses: actions/upload-artifact@v4
