+++ app_front-end/package.json
@@ -21,9 +21,8 @@
     "react-dom": "19.2.0",
     "react-hook-form": "7.71.2",
     "react-icons": "5.5.0",
-    "react-router-dom": "7.13.1",
-    "tailwind-merge": "^3.6.0",
-    "vitest": "4.1.4"
+    "react-router-dom": "^7.17.0",
+    "tailwind-merge": "^3.6.0"
   },
   "devDependencies": {
     "@eslint/js": "9.39.1",
@@ -33,13 +32,17 @@
     "@types/react": "19.2.2",
     "@types/react-dom": "19.2.2",
     "@vitejs/plugin-react": "5.1.2",
+    "@vitest/browser": "^4.1.8",
+    "@vitest/browser-playwright": "^4.1.8",
     "eslint": "9.39.1",
     "eslint-plugin-react-hooks": "7.0.1",
     "eslint-plugin-react-refresh": "0.4.24",
     "globals": "16.5.0",
     "jsdom": "29.0.2",
+    "playwright": "^1.60.0",
     "tailwindcss": "4.3.0",
     "vite": "7.3.3",
-    "vitest": "4.1.4"
+    "vitest": "^4.1.8",
+    "vitest-browser-react": "^2.2.0"
   }
 }
