+++ app_front-end/vite.config.ts
@@ -2,6 +2,7 @@ import path from 'node:path'
 import { defineConfig, loadEnv } from 'vite'
 import react from '@vitejs/plugin-react'
 import tailwindcss from '@tailwindcss/vite'
+import { playwright } from '@vitest/browser-playwright'
 
 // https://vite.dev/config/
 export default defineConfig(({ mode }) => {
@@ -32,7 +33,11 @@ export default defineConfig(({ mode }) => {
         }
       : {}),
     test: {
-      environment: "jsdom",
+      browser: {
+        enabled: true,
+        provider: playwright(),
+        instances: [{ browser: "firefox" }],
+      },
     },
   }
 })
